requirements.txt
# sudo apt-get install libboost-all-dev cmake
# sudo apt-get install libx11-dev
cmake
boost
dlib
albumentations==1.3.0
annoy==1.17.1
blurhash==1.1.4
boost==0.1
certifi==2022.12.7
charset-normalizer==3.0.1
cmake==3.26.0
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
dlib==19.24.0
easydict==1.10
flatbuffers==23.3.3
fonttools==4.39.2
greenlet==2.0.2
humanfriendly==10.0
idna==3.4
imageio==2.26.0
importlib-resources==5.12.0
imutils==0.5.4
inplace-abn==1.1.0
insightface==0.7.2
install==1.3.5
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
Mastodon.py==1.8.0
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
simplejpeg==1.6.5
six==1.16.0
SQLAlchemy==2.0.7
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.3.15
torch @ file:///home/a/torch-1.12.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchaudio @ file:///home/a/torchaudio-0.12.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchdistx @ file:///home/a/torchdistx-0.2.0%2Bcu116-cp38-cp38-manylinux_2_17_x86_64.whl
torchvision @ file:///home/a/torchvision-0.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.15.0
